Move EntityTest into subclasses
By moving all the test methods from EntityTest into subclasses, we make
sure that still all tests are executed when removing the inheritance
between Item/Property and Entity. We can then start to remove tests that
test methods no longer existing while still having the old tests.
